requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
ollama
numpy
passlib[bcrypt]
python-jose
python-multipart
opencv-python
Pillow
